+++ templates/avalonia/CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-_No unreleased changes_
+
+### Changed
+- Update all template target frameworks to .NET 10.
 
 ## [3.0.0-pre13] - 2025-11-19
 ### Added